@@ -9,37 +9,38 @@
     "build": "webpack"
   },
   "devDependencies": {
-    "autoprefixer": "^6.4.0",
-    "babel-core": "^6.14.0",
-    "babel-loader": "^6.2.5",
-    "babel-plugin-transform-decorators-legacy": "^1.3.4",
-    "babel-preset-env": "^1.6.0",
-    "babel-preset-react": "^6.11.1",
-    "babel-preset-stage-1": "^6.13.0",
-    "clean-webpack-plugin": "^0.1.10",
+    "autoprefixer": "^6.7.7",
+    "babel-core": "^6.26.3",
+    "babel-loader": "^6.4.1",
+    "babel-plugin-transform-decorators-legacy": "^1.3.5",
+    "babel-preset-env": "^1.7.0",
+    "babel-preset-react": "^6.24.1",
+    "babel-preset-stage-1": "^6.24.1",
+    "clean-webpack-plugin": "^0.1.19",
     "css-loader": "^0.24.0",
     "file-loader": "^0.9.0",
     "folder-zip": "0.0.5",
     "fs-extra": "^0.30.0",
     "on-build-webpack": "^0.1.0",
-    "postcss-loader": "^0.11.0",
-    "style-loader": "^0.13.1",
-    "webpack": "^1.13.2"
+    "postcss-loader": "^0.11.1",
+    "prettier": "~2.2.1",
+    "style-loader": "^0.13.2",
+    "webpack": "^1.15.0"
   },
   "dependencies": {
-    "classnames": "^2.2.5",
+    "classnames": "^2.2.6",
     "custom-event": "^1.0.1",
     "lodash.debounce": "^4.0.8",
     "lodash.uniq": "^4.5.0",
     "lodash.xor": "^4.5.0",
-    "mobx": "^3.0.2",
-    "mobx-react": "^4.1.0",
-    "qs": "^6.2.1",
-    "react": "^15.4.2",
-    "react-dnd": "^2.1.4",
-    "react-dnd-html5-backend": "^2.1.2",
-    "react-dom": "^15.4.2",
-    "react-lite": "^0.15.19",
-    "react-textarea-autosize": "^4.0.5"
+    "mobx": "^3.6.2",
+    "mobx-react": "^4.4.3",
+    "qs": "^6.9.6",
+    "react": "^15.7.0",
+    "react-dnd": "^2.6.0",
+    "react-dnd-html5-backend": "^2.6.0",
+    "react-dom": "^15.7.0",
+    "react-lite": "^0.15.40",
+    "react-textarea-autosize": "^4.3.2"
   }
 }
