@@ -1,13 +1,15 @@
-import React, {Component} from 'react';
+import React, { Component } from "react";
 
-const iconsSVG = '../images/icons.svg';
+const iconsSVG = "../images/icons.svg";
 
 export default function (props) {
-    const {id} = props;
-    return <svg
-        className={props.className ? props.className : false}
-        onClick={props.onClick ? props.onClick : false}
+  const { id } = props;
+  return (
+    <svg
+      className={props.className ? props.className : false}
+      onClick={props.onClick ? props.onClick : false}
     >
-        <use xlinkHref={iconsSVG + '#' + id}/>
-    </svg>;
+      <use xlinkHref={iconsSVG + "#" + id} />
+    </svg>
+  );
 }
